+++ modules/contracts/apps/challenge/examples/deploy.rs
@@ -4,8 +4,8 @@ use cw_orch::{
     tokio::runtime::Runtime,
 };
 
-use abstract_challenge_app::{contract::CHALLENGE_APP_ID, ChallengeApp};
-use abstract_interface::AppDeployer;
+use abstract_interface::{AppDeployer, DeployStrategy};
+use challenge_app::{contract::CHALLENGE_APP_ID, ChallengeApp};
 use semver::Version;
 
 const CONTRACT_VERSION: &str = env!("CARGO_PKG_VERSION");
@@ -23,6 +23,6 @@ fn main() -> anyhow::Result<()> {
         .build()?;
     let app = ChallengeApp::new(CHALLENGE_APP_ID, chain);
 
-    app.deploy(version)?;
+    app.deploy(version, DeployStrategy::Try)?;
     Ok(())
 }